Y60 now builds under gcc 4.1.1
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@9099 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/components/FFMpegDecoder/FrameAnalyser.cpp
@@ -203,7 +203,7 @@ namespace y60 {
         }
 
 #if LIBAVCODEC_BUILD >= 0x5100
-        int64_t myHalfFrame = av_q2d(_myVStream->time_base) / 2;
+        int64_t myHalfFrame = int64_t(av_q2d(_myVStream->time_base) / 2);
         int64_t myTimestamp = theFrame * (int64_t)av_q2d(_myVStream->time_base) + _myVStream->start_time;
 #else
         int64_t myHalfFrame = int64_t(AV_TIME_BASE) * _myVStream->r_frame_rate_base /
